Fix some address sanitizer bugs
This commit fixes many erroneous malloc-sizeof calls as well as fixing
some types and unused variables. With this commit, the server now
works on macOS.

+++ ircconnpool.h
@@ -9,7 +9,7 @@
 
 struct IrcConnPool;
 
-struct IrcConnPool* ircconnpool_make(int);
+struct IrcConnPool* ircconnpool_make(size_t);
 
 void ircconnpool_free(struct IrcConnPool*);
